requirements.txt
llama-index
pypdf
python-dotenv
llama-index-core
llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-llms-llama-cpp
google-generativeai
llama-index-llms-gemini
streamlit
